+++ library/cartridge_get_control_instance.py
@@ -8,6 +8,7 @@ argument_spec = {
     'play_hosts': {'required': True, 'type': 'list'},
     'console_sock': {'required': True, 'type': 'str'},
     'app_name': {'required': True, 'type': 'str'},
+    'leader_only': {'required': False, 'type': 'bool', 'default': False},
 }
 
 GET_TWOPHASE_COMMIT_VERSION_TIMEOUT = 60
@@ -94,14 +95,26 @@ def candidate_is_ok(uri, names_by_uris, module_hostvars, cluster_disabled_instan
     return helpers.is_enabled(instance_vars)
 
 
-def get_control_instance_name(module_hostvars, cluster_disabled_instances, play_hosts, control_console):
+def get_control_instance_name(
+    module_hostvars,
+    cluster_disabled_instances,
+    play_hosts,
+    control_console,
+    leader_only=False,
+):
     members, err = get_membership_members(control_console)
     if err is not None:
         return None, err
+    leaders, err = helpers.get_active_leaders(control_console)
+    if err is not None:
+        return None, err
+
+    leaders_uuid = set(leaders.values())
 
     alien_members_uris = []
     members_without_uuid = []
     members_by_uuid = {}
+    leaders_uris = []
 
     for uri, member in sorted(members.items()):
         if not member:
@@ -130,6 +143,8 @@ def get_control_instance_name(module_hostvars, cluster_disabled_instances, play_
             continue
 
         members_by_uuid[member_uuid] = member
+        if member_uuid in leaders_uuid:
+            leaders_uris.append(uri)
 
     if alien_members_uris:
         helpers.warn('Incorrect members with the following URIs ignored: %s' % ', '.join(alien_members_uris))
@@ -206,9 +221,17 @@ def get_control_instance_name(module_hostvars, cluster_disabled_instances, play_
     if err is not None:
         return None, "Failed to check instances two-phase commit version: %s" % err
 
-    idx = twophase_commit_versions.index(min(twophase_commit_versions))
-    control_instance_uri = candidates_uris[idx]
+    min_version = min(twophase_commit_versions)
+    min_version_candidates = filter(lambda c: twophase_commit_versions[c[0]] == min_version, enumerate(candidates_uris))
+    candidates_uris = list(map(lambda c: c[1], min_version_candidates))
 
+    if leader_only:
+        leader_candidates_uris = list(filter(lambda uri: uri in leaders_uris, candidates_uris))
+        if len(leader_candidates_uris) == 0:
+            return None, "Not found any leader instance between the candidates: %s" % ', '.join(candidates_uris)
+        candidates_uris = leader_candidates_uris
+
+    control_instance_uri = candidates_uris[0]
     control_instance_name = names_by_uris[control_instance_uri]
 
     return control_instance_name, None
@@ -220,11 +243,12 @@ def get_control_instance(params):
     play_hosts = params['play_hosts']
     console_sock = params['console_sock']
     app_name = params['app_name']
+    leader_only = params.get('leader_only', False)
 
     control_console = helpers.get_control_console(console_sock)
 
     control_instance_name, err = get_control_instance_name(
-        module_hostvars, cluster_disabled_instances, play_hosts, control_console
+        module_hostvars, cluster_disabled_instances, play_hosts, control_console, leader_only
     )
     if err is not None:
         return helpers.ModuleRes(failed=True, msg=err)